Improve performance of storage migration
For #62733

**Problem**
There is a loop that goes through all workspaces in storage (200+ in my case) and then through all keys in storage (5000 in my case). This is quite slow

**optimization**
Remove keys from the set we need to iterate through for each workspace. This cuts the cost of parseStorage in half but it is still obviously quite slow