Do not cache completions that are cancelled during resolve
Part of #63013

**Bug**
When filtering a suggestion list quickly, we try to cancel ongoing resolves. However we cache the resolve too to avoid recomputing them. We currently always cache the result, even if is was canceled. This can result in some suggestion items never being properly resolved

**fix**
On cancellation, invalidate the cached value and make sure we reject the existing promise